feat: parkingzoneData 캐싱 성공

+++ src/main/java/org/example/honorsparkingbe/dto/request/ParkingZoneListRequest.java
@@ -22,5 +22,5 @@ public class ParkingZoneListRequest {
   @Max(value = 180, message = "경도는 180 이하이어야 합니다.")
   private Double longitude;
 
-  private Long page;  // 기본값을 설정
+  private Integer page;  // 기본값을 설정
 }
